requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.34
